+++ src/index.ts
@@ -10,6 +10,7 @@ export { default as Select } from '@/package/components/select.vue';
 export { default as Autocomplete } from '@/package/components/autocomplete.vue';
 export { default as MultiSelect } from '@/package/components/multi-select.vue';
 export { default as DatePicker } from '@/package/components/datepicker.vue';
+export { default as Pagination } from '@/package/components/pagination.vue';
 export { default as Checkbox } from '@/package/components/checkbox.vue';
 export { default as RadioGroup } from '@/package/components/radio-group.vue';
 export { default as RadioButton } from '@/package/components/radio-button.vue';